+++ mocks_test.go
@@ -26,12 +26,12 @@ type mockRecordGetter struct {
 	mock.Mock
 }
 
-func (rg *mockRecordGetter) GetRecords(deviceID string) ([]db.Record, error) {
-	args := rg.Called(deviceID)
+func (rg *mockRecordGetter) GetRecords(deviceID string, limit int) ([]db.Record, error) {
+	args := rg.Called(deviceID, limit)
 	return args.Get(0).([]db.Record), args.Error(1)
 }
 
-func (rg *mockRecordGetter) GetRecordsOfType(deviceID string, eventType int) ([]db.Record, error) {
-	args := rg.Called(deviceID, eventType)
+func (rg *mockRecordGetter) GetRecordsOfType(deviceID string, limit int, eventType int) ([]db.Record, error) {
+	args := rg.Called(deviceID, limit, eventType)
 	return args.Get(0).([]db.Record), args.Error(1)
 }